requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
sentence-transformers
faiss-cpu
pandas
